fix: resolve remaining CI failures
- Fix type mismatch in BlockMemoryPool::AlignBlockSize() for 64-bit platforms
  (Max<T> template deduction failed with uint32_t vs size_t)
- Remove BUILD_C_IFACE=ON from compile-only and test workflows that lack
  STK_C_KERNEL_TYPE_CPU_0 config (cmake-compile-arm-cortex-m, cmake-test-all-x86,
  cmake-test-generic-stm32)

+++ stk/include/memory/stk_memory_blockpool.h
@@ -157,7 +157,7 @@ public:
     */
     static constexpr size_t AlignBlockSize(size_t raw_size)
     {
-        return Max(BLOCK_ALIGN, (raw_size + (BLOCK_ALIGN - 1U)) & ~(BLOCK_ALIGN - 1U));
+        return Max(static_cast<size_t>(BLOCK_ALIGN), (raw_size + (BLOCK_ALIGN - 1U)) & ~static_cast<size_t>(BLOCK_ALIGN - 1U));
     }
 
     /*! \brief     Allocate one block, blocking until one becomes available or the timeout expires.
